# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for planning prestige resets and experiment spending in Necromerger",
        "title": "Necro Prestige Planner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/",
    "paths": {
        "/api/plans": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "List all saved plans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.PlanSummary"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new plan when id is 0 or omitted, updates an existing plan when id \u003e 0",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "Create or update a plan",
                "parameters": [
                    {
                        "description": "Plan data",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/plans/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plans"
                ],
                "summary": "Get a plan by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "plan not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "plans"
                ],
                "summary": "Delete a plan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/recalculate": {
            "post": {
                "description": "Given a plan, returns total Time Shards earned at prestige, experiment costs, rune requirements for planned legendaries, and per-experiment summaries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Calculate Time Shards and resource costs for a plan",
                "parameters": [
                    {
                        "description": "Plan to calculate",
                        "name": "plan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Plan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.RecalculateResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/resource-cap/{threshold}": {
            "post": {
                "description": "Returns current Mana/Slime/Darkness caps, whether the requested Feat threshold is met, and (if not) how many stations of each level are needed per resource to close the gap. Percentage inputs are decimals (0.06 = 6%).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calculator"
                ],
                "summary": "Calculate resource caps and station requirements for a Feat threshold",
                "parameters": [
                    {
                        "enum": [
                            "200k",
                            "400k",
                            "600k",
                            "800k"
                        ],
                        "type": "string",
                        "description": "Combined storage threshold",
                        "name": "threshold",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Player state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ResourceCapRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ResourceCapResponse"
                        }
                    },
                    "400": {
                        "description": "invalid threshold or request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ExperimentSummary": {
            "type": "object",
            "properties": {
                "currentLevel": {
                    "type": "integer"
                },
                "currentLevelCost": {
                    "type": "string"
                },
                "currentLevelValue": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/models.ExperimentID"
                },
                "maxLevel": {
                    "type": "boolean"
                },
                "nextLevelCost": {
                    "type": "string"
                },
                "nextLevelValue": {
                    "type": "string"
                }
            }
        },
        "api.PlanSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "api.RecalculateResponse": {
            "type": "object",
            "properties": {
                "baseShards": {
                    "type": "integer"
                },
                "experimentCost": {
                    "type": "integer"
                },
                "experiments": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/api.ExperimentSummary"
                    }
                },
                "featMultiplier": {
                    "type": "number"
                },
                "legendMultiplier": {
                    "type": "number"
                },
                "legendaryRunes": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/calculator.RuneCosts"
                    }
                },
                "otherMultiplier": {
                    "type": "number"
                },
                "remaining": {
                    "type": "integer"
                },
                "runeNeeded": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "runeTotal": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "totalShards": {
                    "type": "integer"
                }
            }
        },
        "api.RelicLevels": {
            "type": "object",
            "properties": {
                "allResources": {
                    "type": "integer"
                },
                "darkness": {
                    "type": "integer"
                },
                "mana": {
                    "type": "integer"
                },
                "slime": {
                    "type": "integer"
                }
            }
        },
        "api.ResourceCapRequest": {
            "type": "object",
            "properties": {
                "capacityExp1": {
                    "description": "CapacityExp1 is the Capacity Experiment I level (0-9).",
                    "type": "integer"
                },
                "capacityExp2": {
                    "description": "CapacityExp2 is the Capacity Experiment II level (0-9).",
                    "type": "integer"
                },
                "darkStores": {
                    "$ref": "#/definitions/api.StationCounts"
                },
                "fixedTargets": {
                    "description": "FixedTargets lets the player pin a specific cap goal for one resource.\nRemaining combined threshold is then split across the other two.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "goldenBoosts": {
                    "type": "boolean"
                },
                "manaPools": {
                    "$ref": "#/definitions/api.StationCounts"
                },
                "pearlBonus": {
                    "description": "PearlBonus is the per-resource pearl % bonus as shown in game statistics\n(already includes the all-resources pearl contribution).",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "relics": {
                    "$ref": "#/definitions/api.RelicLevels"
                },
                "servO": {
                    "$ref": "#/definitions/api.ServOConfig"
                },
                "skins": {
                    "$ref": "#/definitions/api.ResourceSkins"
                },
                "slimeVats": {
                    "$ref": "#/definitions/api.StationCounts"
                },
                "spells": {
                    "$ref": "#/definitions/api.SpellLevels"
                }
            }
        },
        "api.ResourceCapResponse": {
            "type": "object",
            "properties": {
                "combined": {
                    "type": "integer"
                },
                "darkness": {
                    "type": "integer"
                },
                "delta": {
                    "description": "Delta is positive when the player exceeds the threshold, negative when short.",
                    "type": "integer"
                },
                "gapAnalysis": {
                    "description": "GapAnalysis is only present when the threshold is not met.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/api.ResourceGapAnalysis"
                    }
                },
                "mana": {
                    "type": "integer"
                },
                "met": {
                    "type": "boolean"
                },
                "slime": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
        "api.ResourceGapAnalysis": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "gap": {
                    "type": "integer"
                },
                "options": {
                    "description": "one entry per station level (1-6)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.StationOption"
                    }
                },
                "target": {
                    "type": "integer"
                }
            }
        },
        "api.ResourceSkins": {
            "type": "object",
            "properties": {
                "birthday": {
                    "description": "+5% Slime",
                    "type": "boolean"
                },
                "good": {
                    "description": "+2% All",
                    "type": "boolean"
                },
                "oozing": {
                    "description": "+2000 Slime",
                    "type": "boolean"
                },
                "royalty": {
                    "description": "+5% All",
                    "type": "boolean"
                },
                "santa": {
                    "description": "Multiplicative skins",
                    "type": "boolean"
                },
                "sid": {
                    "description": "+2000 Darkness",
                    "type": "boolean"
                },
                "witch": {
                    "description": "+5% Darkness",
                    "type": "boolean"
                },
                "wizard": {
                    "description": "Base boost skins (flat cap increase)",
                    "type": "boolean"
                }
            }
        },
        "api.ServOConfig": {
            "type": "object",
            "properties": {
                "resource": {
                    "$ref": "#/definitions/models.ResourceType"
                },
                "upgraded": {
                    "type": "boolean"
                }
            }
        },
        "api.SpellLevels": {
            "type": "object",
            "properties": {
                "allResources": {
                    "type": "integer"
                },
                "darkness": {
                    "type": "integer"
                },
                "mana": {
                    "type": "integer"
                },
                "slime": {
                    "type": "integer"
                }
            }
        },
        "api.StationCounts": {
            "type": "object",
            "properties": {
                "l1": {
                    "type": "integer"
                },
                "l2": {
                    "type": "integer"
                },
                "l3": {
                    "type": "integer"
                },
                "l4": {
                    "type": "integer"
                },
                "l5": {
                    "type": "integer"
                },
                "l6": {
                    "type": "integer"
                }
            }
        },
        "api.StationOption": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "runeCost": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "calculator.RuneCosts": {
            "type": "object",
            "additionalProperties": {
                "type": "integer"
            }
        },
        "models.ExperimentID": {
            "type": "string",
            "enum": [
                "seasoning",
                "strength",
                "taste",
                "capacity",
                "body_snatcher",
                "weakening",
                "damage_cap",
                "ice_chest",
                "poison_chest",
                "blood_chest",
                "moon_chest",
                "death_chest",
                "cosmic_chest",
                "seasoning_2",
                "strength_2",
                "taste_2",
                "capacity_2"
            ],
            "x-enum-varnames": [
                "ExpSeasoning",
                "ExpStrength",
                "ExpTaste",
                "ExpCapacity",
                "ExpBodySnatcher",
                "ExpWeakening",
                "ExpDamageCap",
                "ExpIceChest",
                "ExpPoisonChest",
                "ExpBloodChest",
                "ExpMoonChest",
                "ExpDeathChest",
                "ExpCosmicChest",
                "ExpSeasoning2",
                "ExpStrength2",
                "ExpTaste2",
                "ExpCapacity2"
            ]
        },
        "models.Plan": {
            "type": "object",
            "properties": {
                "devourerLevel": {
                    "type": "integer"
                },
                "experimentLevels": {
                    "description": "the planned setup of experiments",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "featTiers": {
                    "type": "integer"
                },
                "groupBonusCount": {
                    "description": "Number of times group bonuses can be claimed (default 1)",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "leftoverShards": {
                    "description": "Shards remaining from previous prestige",
                    "type": "integer"
                },
                "legendaryCounts": {
                    "description": "the planned amount of legendaries",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "otherMultiplier": {
                    "description": "The \"Other\" category (skins, etc.)",
                    "type": "number"
                },
                "possessedLegendaries": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "possessedRunes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.ResourceType": {
            "type": "string",
            "enum": [
                "mana",
                "slime",
                "darkness"
            ],
            "x-enum-varnames": [
                "ResourceMana",
                "ResourceSlime",
                "ResourceDarkness"
            ]
        }
    }
}